+++ CHANGELOG.md
@@ -10,17 +10,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Theme**) Add an option to save the dynamic color theme on the manga page as a custom theme
 
 ### Changed
-- (**Theme**) Change "pure black mode" paper background color to true black
-
-### Changed
 - (**General**) Preserve refresh token (UI Login) over sessions
 - (**Theme**) Improve automatic theme background colors
 - (**Theme**) Adjust "primary color" of dynamic color theme on manga pages
 - (**Theme**) Increase max length of custom theme names to 32 (previously 16)
+- (**Theme**) Change "pure black mode" paper background color to true black
 - (**Library**) Improve library manga sorting
 
 ### Fixed
 - (**Library**) Fix total library size chip color in light mode 
+- (**Browse**) Fix missing pinned sources in the source language filter
 
 ## [20251230.01] (r2937) - 2025-12-30